Separated the data sources and the data into two folders
